Rename composers/ folder to composer/
